# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CounselConnect API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Sign in (student number or staff email)",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Revalidate the current session (CSRF header required)",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "parameters": [
          {
            "name": "counselconnect_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Counselconnect Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Sign out and revoke the session (CSRF header required)",
        "operationId": "logout_api_v1_auth_logout_post",
        "parameters": [
          {
            "name": "counselconnect_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Counselconnect Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Current authenticated user",
        "operationId": "me_api_v1_auth_me_get",
        "parameters": [
          {
            "name": "counselconnect_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Counselconnect Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionUser"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/csrf": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Recover the CSRF token for a live session (safe method, re-issues token)",
        "description": "Re-issues the session-bound CSRF token after a page reload.\n\nSafe because the caller must already hold the session cookie (proof\nof the credential); CSRF tokens are not secrets from their own\nsession — they only guard cross-site forgery of unsafe requests.",
        "operationId": "recover_csrf_api_v1_auth_csrf_get",
        "parameters": [
          {
            "name": "counselconnect_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Counselconnect Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/register/student": {
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Student self-registration (DFD 1.1; account starts PENDING_VERIFICATION)",
        "operationId": "register_student_api_v1_accounts_register_student_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentRegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/register/student-with-cor": {
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Atomic registration: account + COR PDF in ONE submission (DFD 1.1 + 1.2)",
        "operationId": "register_student_with_cor_api_v1_accounts_register_student_with_cor_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_register_student_with_cor_api_v1_accounts_register_student_with_cor_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/SessionUser"
          },
          "csrf_token": {
            "type": "string",
            "title": "Csrf Token"
          },
          "idle_expires_at": {
            "type": "string",
            "title": "Idle Expires At"
          },
          "absolute_expires_at": {
            "type": "string",
            "title": "Absolute Expires At"
          }
        },
        "type": "object",
        "required": [
          "user",
          "csrf_token",
          "idle_expires_at",
          "absolute_expires_at"
        ],
        "title": "AuthResponse"
      },
      "Body_register_student_with_cor_api_v1_accounts_register_student_with_cor_post": {
        "properties": {
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "middle_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Middle Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          },
          "student_number": {
            "type": "string",
            "title": "Student Number"
          },
          "campus_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Campus Id"
          },
          "program_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Program Id"
          },
          "year_level": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Year Level"
          },
          "section": {
            "type": "string",
            "title": "Section"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "Current COR PDF (registration form)"
          }
        },
        "type": "object",
        "required": [
          "first_name",
          "last_name",
          "email",
          "password",
          "student_number",
          "campus_id",
          "program_id",
          "year_level",
          "section",
          "file"
        ],
        "title": "Body_register_student_with_cor_api_v1_accounts_register_student_with_cor_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "identifier": {
            "type": "string",
            "maxLength": 320,
            "minLength": 1,
            "title": "Identifier"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "identifier",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "RegistrationResultResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "profile": {
            "$ref": "#/components/schemas/StudentProfileResponse"
          },
          "next_step": {
            "type": "string",
            "title": "Next Step"
          }
        },
        "type": "object",
        "required": [
          "user",
          "profile",
          "next_step"
        ],
        "title": "RegistrationResultResponse",
        "description": "Result of student self-registration (always PENDING_VERIFICATION)."
      },
...
      "SessionUser": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role_code": {
            "type": "string",
            "title": "Role Code"
          },
          "account_status": {
            "type": "string",
            "title": "Account Status"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "email",
          "role_code",
          "account_status",
          "first_name",
          "last_name"
        ],
        "title": "SessionUser"
      },
      "StudentProfileResponse": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "student_number": {
            "type": "string",
            "title": "Student Number"
          },
          "campus_id": {
            "type": "integer",
            "title": "Campus Id"
          },
          "program_id": {
            "type": "integer",
            "title": "Program Id"
          },
          "year_level": {
            "type": "integer",
            "title": "Year Level"
          },
          "section": {
            "type": "string",
            "title": "Section"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "student_number",
          "campus_id",
          "program_id",
          "year_level",
          "section"
        ],
        "title": "StudentProfileResponse"
      },
      "StudentRegistrationRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          },
          "first_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "First Name"
          },
          "middle_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Middle Name"
          },
          "last_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Last Name"
          },
          "student_number": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Student Number"
          },
          "campus_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Campus Id"
          },
          "program_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Program Id"
          },
          "year_level": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Year Level"
          },
          "section": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Section"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password",
          "first_name",
          "last_name",
          "student_number",
          "campus_id",
          "program_id",
          "year_level",
          "section"
        ],
        "title": "StudentRegistrationRequest",
        "description": "DFD 1.1 'Register Student Account' payload.\n\nCreates the pending account + academic profile; COR upload is a\nseparate step (limits pending ADR-P05)."
      },
...
      "UserResponse": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "role_code": {
            "type": "string",
            "enum": [
              "STUDENT",
              "GUIDANCE_STAFF",
              "COUNSELOR"
            ],
            "title": "Role Code"
          },
          "account_status": {
            "type": "string",
            "enum": [
              "PENDING_VERIFICATION",
              "ACTIVE",
              "VERIFICATION_EXPIRED"
            ],
            "title": "Account Status"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "middle_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Middle Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "email",
          "role_code",
          "account_status",
          "first_name",
          "middle_name",
          "last_name",
          "created_at",
          "updated_at"
        ],
        "title": "UserResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}